Rename ReflectionException to DeserializationException.
