+++ CMakeLists.txt
@@ -494,8 +494,7 @@ endif()
 # --------------------------------------------------------
 # Serialization extensions
 # --------------------------------------------------------
-add_subdirectory(serialization/common/base)
-add_subdirectory(serialization/common/core)
+add_subdirectory(serialization/common/common)
 
 # Capnroto
 if (ECAL_USE_CAPNPROTO)
