+++ serialization/common/common/CMakeLists.txt
@@ -16,9 +16,6 @@
 #
 # ========================= eCAL LICENSE =================================
 
-################################
-# eCAL core Message extensions #
-################################
 add_library(message_base INTERFACE)
 add_library(eCAL::message_base ALIAS message_base)
 
@@ -28,7 +25,7 @@ target_sources(message_base
     TYPE HEADERS
     BASE_DIRS include
     FILES
-      include/ecal/msg/common/exception.h
+      include/ecal/msg/exception.h
 )
 
 target_compile_features(message_base INTERFACE cxx_std_14)
@@ -37,4 +34,37 @@ install(
   TARGETS message_base
   EXPORT eCALCoreTargets  
   FILE_SET message_base_headers COMPONENT sdk
+)
+
+
+################################
+# eCAL core Message extensions #
+################################
+add_library(message_core INTERFACE)
+add_library(eCAL::message_core ALIAS message_core)
+
+target_link_libraries(message_core 
+  INTERFACE 
+    eCAL::core
+)
+
+target_sources(message_core
+  INTERFACE
+    FILE_SET message_core_headers
+    TYPE HEADERS
+    BASE_DIRS include
+    FILES
+      include/ecal/msg/dynamic.h
+      include/ecal/msg/publisher.h
+      include/ecal/msg/subscriber.h
+)
+
+target_compile_features(message_core INTERFACE cxx_std_14)
+
+install(
+  TARGETS message_core
+  EXPORT eCALCoreTargets  
+  ARCHIVE       DESTINATION "${eCAL_install_archive_dir}" COMPONENT sdk
+  LIBRARY       DESTINATION "${eCAL_install_lib_dir}"     COMPONENT sdk
+  FILE_SET message_core_headers COMPONENT sdk
 )
