+++ serialization/common/common/include/ecal/msg/dynamic.h
@@ -29,6 +29,7 @@
 #include <mutex>
 #include <string>
 
+#include <ecal/msg/exception.h>
 #include <ecal/pubsub/subscriber.h>
 #include <ecal/util.h>
 
@@ -223,7 +224,7 @@ namespace eCAL
         auto msg = m_deserializer.Deserialize(data_.buffer, data_.buffer_size, topic_info_);
         fn_callback(topic_id_, msg, data_.send_timestamp, data_.send_clock);
       }
-      catch (const DynamicReflectionException& e)
+      catch (const DeserializationException& e)
       {
         CallErrorCallback(std::string("Dynamic Deserialization: Error deserializing data: ") + e.what() );
       }
