+++ serialization/common/common/include/ecal/msg/message.h
