+++ serialization/common/common/include/ecal/msg/publisher.h
