+++ serialization/common/common/include/ecal/msg/subscriber.h
